requirements.txt
pdfminer.six
sentence-transformers
text-generation
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask==2.0.3
Werkzeug==2.0.3
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn[standard]>=0.15.0,<0.16.0
einops